initial move of loc importer to new module
