+++ ohmg/loc_insurancemaps/models.py
@@ -30,42 +30,13 @@ from ohmg.georeference.models import (
 )
 from ohmg.georeference.storage import OverwriteStorage
 from ohmg.places.models import Place
-
-from ohmg.loc_insurancemaps.utils import (
-    LOCParser,
-    LOCConnection,
-    filter_volumes_for_use,
-    unsanitize_name,
-    get_jpg_from_jp2_url,
-)
 from ohmg.core.utils import (
+    get_jpg_from_jp2_url,
     STATE_CHOICES,
     STATE_ABBREV,
     MONTH_CHOICES,
 )
 logger = logging.getLogger(__name__)
-
-def write_trim_feature_cache(feature, file_path):
-    with open(file_path, "w") as f:
-        json.dump(feature, f, indent=2)
-
-def read_trim_feature_cache(file_path):
-    with open(file_path, "r") as f:
-        feature = json.load(f)
-    return feature
-
-def check_trim_feature_cache(in_path, feature):
-
-    feat_cache_path = in_path.replace(".tif", "_trim-feature.json")
-    if os.path.isfile(feat_cache_path):
-        cached_feature = read_trim_feature_cache(feat_cache_path)
-    else:
-        cached_feature = None
-        write_trim_feature_cache(feature, feat_cache_path)
-
-    return cached_feature
-    
-
 
 def find_volume(item):
     """Attempt to get the volume from which a Document or Layer
@@ -416,6 +387,8 @@ class Volume(models.Model):
 
     def make_sheets(self):
 
+        from ohmg.core.importers.loc_sanborn import LOCParser
+
         files_to_import = self.lc_resources[0]['files']
         for fileset in files_to_import:
             parsed = LOCParser(fileset=fileset)
@@ -714,71 +687,6 @@ class Volume(models.Model):
             data['sessions'] = self.get_user_activity_summary()
 
         return data
-
-    def import_all_available_volumes(self, state, apply_filter=True, verbose=False):
-        """Preparatory step that runs through all cities in the provided
-        state, filters the available volumes for those cities, and then
-        imports each one to create a new Volume object."""
-
-        lc = LOCConnection(delay=0, verbose=verbose)
-        cities = lc.get_city_list_by_state(state)
-
-        volumes = []
-        for city in cities:
-            lc.reset()
-            city = unsanitize_name(state, city[0])
-            vols = lc.get_volume_list_by_city(city, state)
-            if apply_filter is True:
-                vols = filter_volumes_for_use(vols)
-                volumes += [i for i in vols if i['include'] is True]
-            else:
-                volumes += vols
-
-        for volume in volumes:
-            try:
-                Volume.objects.get(pk=volume['identifier'])
-            except Volume.DoesNotExist:
-                Volume().import_volume(volume['identifier'])
-
-    def import_volume(self, identifier, locale=None, dry_run=False):
-
-        try:
-            volume = Volume.objects.get(pk=identifier)
-            volume.locales.set([locale])
-            volume.update_place_counts()
-
-            # make sure a main-content layerset exists for this volume
-            main_ls, _ = AnnotationSet.objects.get_or_create(
-                category=SetCategory.objects.get(slug="main-content"),
-                volume=volume,
-            )
-            return volume
-        except Volume.DoesNotExist:
-            pass
-
-        lc = LOCConnection(delay=0, verbose=True)
-        response = lc.get_item(identifier)
-        if response.get("status") == 404:
-            return None
-
-        parsed = LOCParser(item=response['item'])
-        volume_kwargs = parsed.volume_kwargs()
-
-        # add resources to args, not in item (they exist adjacent)
-        volume_kwargs["lc_resources"] = response['resources']
-
-        if dry_run:
-            return volume_kwargs
-        else:
-            volume = Volume.objects.create(**volume_kwargs)
-            volume.locales.add(locale)
-            volume.update_place_counts()
-            # make sure a main-content layerset exists for this volume
-            main_ls, _ = AnnotationSet.objects.get_or_create(
-                category=SetCategory.objects.get(slug="main-content"),
-                volume=volume,
-            )
-            return volume
 
 ## This seems to be where the signals need to be connected. See
 ## https://github.com/mradamcox/loc-insurancemaps/issues/75
